branches/20190902_v1_1/src/nbr/src/test/java/com/bx/erp/model/VipConsumeHistoryTest.java
package com.bx.erp.model;

import org.testng.Assert;
import org.testng.annotations.AfterClass;
import org.testng.annotations.BeforeClass;
import org.testng.annotations.Test;

import com.bx.erp.action.bo.BaseBO;
import com.bx.erp.test.Shared;

public class VipConsumeHistoryTest {

	@BeforeClass
	public void setUp() {
		Shared.printTestClassStartInfo();
	}

	@AfterClass
	public void tearDown() {
		Shared.printTestClassEndInfo();
	}

	@Test
	public void testCheckCreate() {
		Shared.printTestMethodStartInfo();
		
		VipConsumeHistory vipConsumeHistory = new VipConsumeHistory();
		String error = "";
		error = vipConsumeHistory.checkCreate(BaseBO.INVALID_CASE_ID);
		Assert.assertEquals(error, "");
	}

	@Test
	public void testCheckRetrieve1() {
		Shared.printTestMethodStartInfo();
		
		VipConsumeHistory vipConsumeHistory = new VipConsumeHistory();
		String error = "";
		error = vipConsumeHistory.checkRetrieve1(BaseBO.INVALID_CASE_ID);
		Assert.assertEquals(error, "");
	}

	@Test
	public void testCheckRetrieveN() {
		Shared.printTestMethodStartInfo();
		
		VipConsumeHistory vipConsumeHistory = new VipConsumeHistory();
		String error = "";
		error = vipConsumeHistory.checkRetrieveN(BaseBO.INVALID_CASE_ID);
		Assert.assertEquals(error, "");
	}

	@Test
	public void testCheckUpdate() {
		Shared.printTestMethodStartInfo();
		
		VipConsumeHistory vipConsumeHistory = new VipConsumeHistory();
		String error = "";
		error = vipConsumeHistory.checkUpdate(BaseBO.INVALID_CASE_ID);
		Assert.assertEquals(error, "");
	}

	@Test
	public void testCheckDelete() {
		Shared.printTestMethodStartInfo();
		
		VipConsumeHistory vipConsumeHistory = new VipConsumeHistory();
		String error = "";
		error = vipConsumeHistory.checkDelete(BaseBO.INVALID_CASE_ID);
		Assert.assertEquals(error, "");
	}
}
